requirements.txt
Flask
google-generativeai
matplotlib
textblob
speechrecognition
python-dotenv
